breaking change in behavior for tasks that build to $2
Previously a result in $2 that hasn't changed (via `touch`)
would be deleted if its a file, but left alone if its a dir.

Now, such results are deleted in all cases, unless they are a symbolic
link (because it's impossible tou `touch` a symlink on most OSes).

If you have build targets that create a directory at "$2", you'll
need to make sure you `touch "$2"` if it already exists.